#237 - Move @Query annotation to o.s.d.r2dbc.repository.
