+++ src/main/java/org/springframework/data/r2dbc/repository/query/R2dbcQueryMethod.java
@@ -29,6 +29,7 @@ import org.springframework.data.domain.Sort;
 import org.springframework.data.mapping.context.MappingContext;
 import org.springframework.data.projection.ProjectionFactory;
 import org.springframework.data.r2dbc.repository.Modifying;
+import org.springframework.data.r2dbc.repository.Query;
 import org.springframework.data.relational.core.mapping.RelationalPersistentEntity;
 import org.springframework.data.relational.core.mapping.RelationalPersistentProperty;
 import org.springframework.data.relational.repository.query.RelationalEntityMetadata;
@@ -60,7 +61,7 @@ public class R2dbcQueryMethod extends QueryMethod {
 
 	private final Method method;
 	private final MappingContext<? extends RelationalPersistentEntity<?>, ? extends RelationalPersistentProperty> mappingContext;
-	private final Optional<Query> query;
+	private final Optional<org.springframework.data.r2dbc.repository.Query> query;
 	private final boolean modifying;
 
 	private @Nullable RelationalEntityMetadata<?> metadata;
@@ -110,7 +111,8 @@ public class R2dbcQueryMethod extends QueryMethod {
 		}
 
 		this.method = method;
-		this.query = Optional.ofNullable(AnnotatedElementUtils.findMergedAnnotation(method, Query.class));
+		this.query = Optional.ofNullable(
+				AnnotatedElementUtils.findMergedAnnotation(method, org.springframework.data.r2dbc.repository.Query.class));
 		this.modifying = AnnotatedElementUtils.hasAnnotation(method, Modifying.class);
 	}
 
@@ -207,14 +209,14 @@ public class R2dbcQueryMethod extends QueryMethod {
 	}
 
 	/**
-	 * Returns the required query string declared in a {@link Query} annotation or throws {@link IllegalStateException} if
-	 * neither the annotation found nor the attribute was specified.
+	 * Returns the required query string declared in a {@link org.springframework.data.r2dbc.repository.Query} annotation
+	 * or throws {@link IllegalStateException} if neither the annotation found nor the attribute was specified.
 	 *
 	 * @return the query string.
 	 * @throws IllegalStateException in case query method has no annotated query.
 	 */
 	public String getRequiredAnnotatedQuery() {
-		return this.query.map(Query::value)
+		return this.query.map(org.springframework.data.r2dbc.repository.Query::value)
 				.orElseThrow(() -> new IllegalStateException("Query method " + this + " has no annotated query"));
 	}
 
@@ -223,12 +225,13 @@ public class R2dbcQueryMethod extends QueryMethod {
 	 *
 	 * @return the optional query annotation.
 	 */
-	Optional<Query> getQueryAnnotation() {
+	Optional<org.springframework.data.r2dbc.repository.Query> getQueryAnnotation() {
 		return this.query;
 	}
 
 	/**
-	 * @return {@literal true} if the {@link Method} is annotated with {@link Query}.
+	 * @return {@literal true} if the {@link Method} is annotated with
+	 *         {@link org.springframework.data.r2dbc.repository.Query}.
 	 */
 	public boolean hasAnnotatedQuery() {
 		return getQueryAnnotation().isPresent();
